auth_methods :dizzy: all the methods are implimneted and all in place tested most of them :dizzy: :/

+++ rest_api/views.py
@@ -7,7 +7,7 @@ from rest_framework.decorators import (authentication_classes, permission_classe
 from rest_framework_api_key.permissions import HasAPIKey
 
 from utils.util import compose_into_a_single_decorator
-from school_management_system.authentications import authenticate_user
+from school_management_system.authentications import (authenticate_user, is_token_authenticated, authenticate_api_key)
 from rest_api.models import (StudentDetails, TeacherDetails, ParentsDetails, DriverDetails)
 from rest_api.serializers import (StudentDetailsSerializer, TeacherDetailsSerializer, ParentsDetailsSerializer, DriverDetailsSerializer)
 
@@ -17,15 +17,15 @@ not_use_token_as_default_auth = compose_into_a_single_decorator(authentication_c
 
 class CreateStudent(APIView):
 
-    permission_classes = (IsAuthenticated,)
-
     def post(self, request):
-        data = JSONParser().parse(request)
-        tutorial_serializer = StudentDetailsSerializer(data=data)
-        if tutorial_serializer.is_valid():
-            tutorial_serializer.save()
-            return JsonResponse(tutorial_serializer.data, status=status.HTTP_201_CREATED)
-        return JsonResponse(tutorial_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if is_token_authenticated(request):
+            data = JSONParser().parse(request)
+            tutorial_serializer = StudentDetailsSerializer(data=data)
+            if tutorial_serializer.is_valid():
+                tutorial_serializer.save()
+                return JsonResponse(tutorial_serializer.data, status=status.HTTP_201_CREATED)
+            return JsonResponse(tutorial_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return JsonResponse({'error': 'key is not valid'}, status=status.HTTP_401_UNAUTHORIZED)
 
     def put(self, request):
         return self.post(request)
@@ -36,15 +36,15 @@ class SearchStudent(APIView):
     ?name=<value>
     """
 
-    permission_classes = (IsAuthenticated,)
-
     def get(self, request):
-        tutorials = StudentDetails.objects.all()
-        name = request.query_params.get('name', None)
-        if name is not None:
-            tutorials = StudentDetails.objects.filter(student_name__icontains=name)
-        tutorials_serializer = StudentDetailsSerializer(tutorials, many=True)
-        return JsonResponse(tutorials_serializer.data, safe=False)
+        if is_token_authenticated(request):
+            tutorials = StudentDetails.objects.all()
+            name = request.query_params.get('name', None)
+            if name is not None:
+                tutorials = StudentDetails.objects.filter(student_name__icontains=name)
+            tutorials_serializer = StudentDetailsSerializer(tutorials, many=True)
+            return JsonResponse(tutorials_serializer.data, safe=False)
+        return JsonResponse({'error': 'key is not valid'}, status=status.HTTP_401_UNAUTHORIZED)
 
 
 class DeleteStudent(APIView):
@@ -52,16 +52,17 @@ class DeleteStudent(APIView):
     permission_classes = (IsAuthenticated,)
 
     def delete(self, request):
-        name = request.query_params.get('name', None)
-        if name:
-            delete_value = StudentDetails.objects.filter(student_name=name).delete()
-            if delete_value[0]:
-                return JsonResponse({'message': 'Tutorial was deleted successfully!'}, status=status.HTTP_204_NO_CONTENT)
+        if is_token_authenticated(request):
+            name = request.query_params.get('name', None)
+            if name:
+                delete_value = StudentDetails.objects.filter(student_name=name).delete()
+                if delete_value[0]:
+                    return JsonResponse({'message': 'Tutorial was deleted successfully!'}, status=status.HTTP_204_NO_CONTENT)
+                return JsonResponse({'message': 'The tutorial does not exist'}, status=status.HTTP_404_NOT_FOUND)
             return JsonResponse({'message': 'The tutorial does not exist'}, status=status.HTTP_404_NOT_FOUND)
-        return JsonResponse({'message': 'The tutorial does not exist'}, status=status.HTTP_404_NOT_FOUND)
+        return JsonResponse({'error': 'key is not valid'}, status=status.HTTP_401_UNAUTHORIZED)
 
 
-@not_use_token_as_default_auth
 class CreateParent(APIView):
 
     def post(self, request, *args, **kwargs):
@@ -78,7 +79,6 @@ class CreateParent(APIView):
         return self.post(request)
 
 
-@not_use_token_as_default_auth
 class SearchParent(APIView):
 
     def get(self, request):
@@ -93,7 +93,6 @@ class SearchParent(APIView):
         return JsonResponse({'error': True}, status=status.HTTP_401_UNAUTHORIZED)
 
 
-@not_use_token_as_default_auth
 class DeleteParent(APIView):
 
     def delete(self, request):
@@ -112,12 +111,14 @@ class CreateTeacher(APIView):
     permission_classes = (HasAPIKey,)
 
     def post(self, request, *args, **kwargs):
-        tutorial_data = JSONParser().parse(request)
-        tutorial_serializer = TeacherDetailsSerializer(data=tutorial_data)
-        if tutorial_serializer.is_valid():
-            tutorial_serializer.save()
-            return JsonResponse(tutorial_serializer.data, status=status.HTTP_201_CREATED)
-        return JsonResponse(tutorial_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if authenticate_api_key(request):
+            tutorial_data = JSONParser().parse(request)
+            tutorial_serializer = TeacherDetailsSerializer(data=tutorial_data)
+            if tutorial_serializer.is_valid():
+                tutorial_serializer.save()
+                return JsonResponse(tutorial_serializer.data, status=status.HTTP_201_CREATED)
+            return JsonResponse(tutorial_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return JsonResponse({'error': True}, status=status.HTTP_401_UNAUTHORIZED)
 
     def put(self, request):
         return self.post(request)
@@ -128,33 +129,33 @@ class SearchTeacher(APIView):
     permission_classes = (HasAPIKey, )
 
     def get(self, request):
+        if authenticate_api_key(request):
+            tutorials = TeacherDetails.objects.all()
+            name = request.query_params.get('name', None)
+            if name is not None:
+                tutorials = TeacherDetails.objects.filter(teacher_name__icontains=name)
 
-        tutorials = TeacherDetails.objects.all()
-        name = request.query_params.get('name', None)
-        if name is not None:
-            tutorials = TeacherDetails.objects.filter(teacher_name__icontains=name)
-
-        tutorials_serializer = TeacherDetailsSerializer(tutorials, many=True)
-        return JsonResponse(tutorials_serializer.data, safe=False)
+            tutorials_serializer = TeacherDetailsSerializer(tutorials, many=True)
+            return JsonResponse(tutorials_serializer.data, safe=False)
+        return JsonResponse({'error': True}, status=status.HTTP_401_UNAUTHORIZED)
 
 
-@not_use_token_as_default_auth
 class DeleteTeacher(APIView):
 
     permission_classes = (HasAPIKey,)
 
     def delete(self, request):
-
-        name = request.query_params.get('name', None)
-        if name:
-            delete_value = TeacherDetails.objects.filter(teacher_name=name).delete()
-            if delete_value[0]:
-                return JsonResponse({'message': 'Tutorial was deleted successfully!'}, status=status.HTTP_204_NO_CONTENT)
+        if authenticate_api_key(request):
+            name = request.query_params.get('name', None)
+            if name:
+                delete_value = TeacherDetails.objects.filter(teacher_name=name).delete()
+                if delete_value[0]:
+                    return JsonResponse({'message': 'Tutorial was deleted successfully!'}, status=status.HTTP_204_NO_CONTENT)
+                return JsonResponse({'message': 'The tutorial does not exist'}, status=status.HTTP_404_NOT_FOUND)
             return JsonResponse({'message': 'The tutorial does not exist'}, status=status.HTTP_404_NOT_FOUND)
-        return JsonResponse({'message': 'The tutorial does not exist'}, status=status.HTTP_404_NOT_FOUND)
+        return JsonResponse({'error': True}, status=status.HTTP_401_UNAUTHORIZED)
 
 
-@not_use_token_as_default_auth
 class CreateDriver(APIView):
 
     def post(self, request, *args, **kwargs):
@@ -169,7 +170,6 @@ class CreateDriver(APIView):
         return self.post(request)
 
 
-@not_use_token_as_default_auth
 class SearchDriver(APIView):
 
     def get(self, request, *args, **kwargs):
@@ -182,7 +182,6 @@ class SearchDriver(APIView):
         return JsonResponse(tutorials_serializer.data, safe=False)
 
 
-@not_use_token_as_default_auth
 class DeleteDriver(APIView):
 
     def delete(self, request, *args, **kwargs):
